Fixed some routes and functions
