+++ lib/functions.js
@@ -37,4 +37,15 @@ async function LoggedIn(req, res, next) {
         res.redirect("/");
     }
 }
-module.exports = { LoggedInPost, log, error, debug, LoggedIn };
+function CheckUser(user) {
+    return {
+        discord_user_id: user.user?.discord_user_id,
+        email: user.user.email,
+        id: user.user.id,
+        profile_pic: user.user.profile_pic,
+        username: user.user.username,
+        account: user.account,
+        channel_link: user.channel_link
+    };
+}
+module.exports = { LoggedInPost, log, error, debug, LoggedIn, CheckUser };
