+++ routes/admin.js
@@ -4,16 +4,6 @@ var fs = require("fs");
 var path = require("path");
 const functions = require("../lib/functions");
 const DataBase = require("../model/DataBase");
-function CheckUser(user) {
-    return {
-        discord_user_id: user.user.discord_user_id,
-        email: user.user.email,
-        id: user.user.id,
-        profile_pic: user.user.profile_pic,
-        username: user.user.username,
-        account: user.account
-    };
-}
 function isAdmin(user) {
     if (user.account.type == "staff") return true | false;
 }
@@ -21,7 +11,7 @@ app.get("/dashboard/staff", functions.LoggedIn, async (req, res) => {
     const user = await DataBase.findOne({ "user.id": req.session.user.user.id }).exec();
     if (isAdmin(user)) {
         res.render(path.resolve("./views/admin/panel.ejs"), {
-            user: CheckUser(user)
+            user: functions.CheckUser(user)
         });
     } else {
         res.redirect("/dashboard");
@@ -31,7 +21,7 @@ app.get("/dashboard/staff/rewards_and_users/:slug", functions.LoggedIn, async (r
     const user = await DataBase.findOne({ "user.id": req.session.user.user.id }).exec();
     if (isAdmin(user)) {
         res.render(path.resolve("./views/admin/rewards_and_users.ejs"), {
-            user: CheckUser(user),
+            user: functions.CheckUser(user),
             slug: req.params.slug
         });
     } else {
